Extend Shields context interface + split existing tests
